Querying + Tests

+++ src/main/java/org/musicshop/pojo/Relationship.java
@@ -4,6 +4,6 @@ public interface Relationship extends Product {
 
     long getRelated();
     void setRelated(long related);
-    long getM1();
-    void setM1(long m1);
+    double getM1();
+    void setM1(double m1);
 }
